+++ ai-terminal/src-tauri/src/command/core/mod.rs
@@ -1,1 +1,2 @@
 pub mod execute_command;
+pub mod terminate_command;
