+++ ai-terminal/src-tauri/src/command/mod.rs
@@ -1,4 +1,3 @@
 pub mod core;
 pub mod git_commands;
 pub mod types;
-pub mod utils;
