--- ai-terminal/src-tauri/src/command/utils/mod.rs
@@ -1,1 +1,0 @@
-pub mod path_utils;
