+++ ai-terminal/src-tauri/src/utils/file_system_utils.rs
