+++ ai-terminal/src-tauri/src/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod file_system_utils;
 pub mod operating_system_utils;
