+++ ai-terminal/src-tauri/src/utils/operating_system_utils.rs
@@ -12,3 +12,9 @@ pub fn get_operating_system() -> String {
     #[cfg(not(any(target_os = "windows", target_os = "macos", target_os = "linux")))]
     return "Unknown".to_string();
 }
+
+#[tauri::command]
+pub fn get_system_environment_variables() -> Result<Vec<(String, String)>, String> {
+    let env_vars: Vec<(String, String)> = std::env::vars().collect();
+    Ok(env_vars)
+}
